requirements.txt
psycopg2-binary
infi.clickhouse-orm
boto3